Add fixed price of 5 to missing dates
